Update TP 1.

+++ TP-1/ej-3/matrix_without_threads.cpp
@@ -1,13 +1,12 @@
 #include <iostream>
 #include <vector>
 #include <chrono>
-#include <thread>
 
 using namespace std;
 
 // multiplicación de dos matrices de punto flotante
 vector<vector<float>> multiply_matrices(const vector<vector<float>>& matrix1, const vector<vector<float>>& matrix2) {
-    int N = matrix1.size();
+    int N = (int) matrix1.size();
     vector<vector<float>> result(N, vector<float>(N, 0.0));  // inicializa la matriz de resultados con ceros
 
     for (int i = 0; i < N; i++) {
@@ -23,7 +22,7 @@ vector<vector<float>> multiply_matrices(const vector<vector<float>>& matrix1, co
 
 // calcula la sumatoria de todos los elementos de una matriz
 float sum_elements(const vector<vector<float>>& matrix) {
-    int N = matrix.size();
+    int N = (int) matrix.size();
     float sum = 0.0;
 
     for (int i = 0; i < N; i++) {
@@ -36,7 +35,7 @@ float sum_elements(const vector<vector<float>>& matrix) {
 }
 
 chrono::duration<double> matrix_without_threads(const vector<vector<float>>& matrix1, const vector<vector<float>>& matrix2) {
-    int N = matrix1.size();
+    int N = (int) matrix1.size();
     
     cout << "Sin Hilos:" << endl;
     cout << "Calculando..." << endl;
